+++ src/main/java/dndbuilder/characterbuilder/CharacterClass.java
@@ -1,4 +1,4 @@
-package characterFiles;
+package characterbuilder;
 
 /**
  * Version: 0.1.Alpha 2020/08/27 Steven Anderson