+++ src/main/java/dndbuilder/spellbuilder/Spell.java
@@ -1,4 +1,4 @@
-package characterFiles;
+package spellbuilder;
 
 /**
  * Version: 0.1.Alpha 2020/08/27 Steven Anderson