+++ src/main/java/dndbuilder/spellbuilder/SpellScribe.java
@@ -1,0 +1,15 @@
+package spellbuilder;
+
+/**
+ * Version: 0.1.Alpha 2020/08/30 Steven Anderson
+ * This is a shared program to build spells for the game. This part of the program can be run separately from the character creation.
+ */
+public class SpellScribe {
+    
+    public SpellScribe(){
+        //TODO: 
+        System.out.println("Create spells for Dungeons and Dragons 5e");
+
+    }
+    
+}